feat: airbyte pipeline (pg cdc -> airbyte -> mssql) via abctl + public API
Task 5. Adds airbyte/configure.py (public-API client: OAuth bearer token,
source/destination/connection create, sync driver) and make airbyte-install/
-up/-down/-bench.

MEASURED, real number (RATE=20 DURATION=60 MIX=100/0/0 GRACE=180):
  completeness 100% (1156/1156)
  p50 15787 ms | p95 37663 ms | p99 40077 ms | max 40694 ms
  throughput 19.3 rows/s
vs Debezium p50 ~520-645 ms on the same rig — the streaming-vs-batch gap,
measured end to end. Airbyte polls a batch sync; a row's latency is dominated by
how long until the next sync picks it up.

Getting here required working around several real integration frictions, all now
encoded in the Makefile/configure.py so the documented commands reproduce:
- abctl official installer returns Cloudflare 526 -> airbyte-install falls back
  to the pinned GitHub release (v0.30.4).
- ingress port 8000 clashed with a local dev API -> default AIRBYTE_PORT=8010.
- abctl gates the internal Config API behind session auth -> configure.py uses
  the PUBLIC API (/api/public/v1) with an app client-id/secret bearer token.
- Airbyte's connector pods run in abctl's kind cluster, a separate docker
  network -> airbyte-up connects the DB containers to the kind network and
  passes their kind-network IPs as AB_SRC_HOST/AB_DST_HOST (compose service
  names don't resolve there; symptom was JDBC 08001).
- source-postgres CDC requires initial_waiting_seconds >= 120 (watchdog timeout
  otherwise).
- destination-mssql 2.2.20 crashes introspecting a pre-created table whose
  column types aren't in its MssqlType enum (FLOAT, NVARCHAR) -> the connector
  must OWN its target table, so airbyte-up drops dbo.source_events first.
- schema: written_at changed DOUBLE PRECISION -> BIGINT (epoch micros) across
  sql/ + harness, since the connector cannot map FLOAT. Latency is timed on the
  host clock, not this column, so the change is measurement-neutral.
- Airbyte job status "incomplete" is transient (auto-retry can still reach
  "succeeded") -> sync() waits only on succeeded/failed/cancelled, bounded.
- airbyte-bench sync loop now runs for DURATION+GRACE so the batch tail arrives
  during the drain; --grace is passed to the harness (GRACE make var, default 30).

Note: the two tools need different target-table shapes (harness-created for
Debezium, Airbyte-created for Airbyte), so `make up` alone does not restore a
clean rig between tools — switching tools requires the tool's own -up target.

+++ sql/mssql_init.sql
@@ -8,7 +8,7 @@ IF OBJECT_ID('dbo.source_events','U') IS NULL
 CREATE TABLE dbo.source_events (
     id          BIGINT NOT NULL PRIMARY KEY,
     seq         BIGINT NOT NULL,
-    written_at  FLOAT  NOT NULL,
+    written_at  BIGINT NOT NULL,
     payload     NVARCHAR(MAX) NULL
 );
 GO
